# Conflicts:
# Entities/Sound.cpp - deleted and updated my renamed version
# Entities/Sound.h - deleted and updated my renamed version
# Entities/ThrownDevice.h - resolved manually
# GUI/GUISound.h - used mine then manually changed
# Managers/AudioMan.cpp - used mine then manually changed
# Managers/AudioMan.h - used mine then manually changed
# RTEA.vcxproj - resolved manually
# System/ContentFile.cpp - used theirs then manually changed
# System/ContentFile.h - used theirs then manually changed
# external/lib/osx/libfmod.a - used mine

+++ GUI/GUIFont.h
@@ -52,7 +52,7 @@ public:
 
     // Font Color structure
     typedef struct {
-        Uint32        m_Color;
+        unsigned long        m_Color;
         GUIBitmap    *m_Bitmap;
     } FontColor;
 
@@ -82,7 +82,7 @@ public:
 // Description:     Pre-Calculates the font using a specific color.
 // Arguments:       Color.
 
-    void CacheColor(Uint32 Color);
+    void CacheColor(unsigned long Color);
 
 
 //////////////////////////////////////////////////////////////////////////////////////////
@@ -91,7 +91,7 @@ public:
 // Description:     Finds a font color structure from the cache.
 // Arguments:       Color.
 
-    FontColor *GetFontColor(Uint32 Color);
+    FontColor *GetFontColor(unsigned long Color);
 
 
 //////////////////////////////////////////////////////////////////////////////////////////
@@ -100,7 +100,7 @@ public:
 // Description:     Draws text to a bitmap.
 // Arguments:       Bitmap, Position, Text, Color, Drop-shadow, 0 = none.
 
-    void Draw(GUIBitmap *Bitmap, int X, int Y, const std::string Text, Uint32 Shadow = 0);
+    void Draw(GUIBitmap *Bitmap, int X, int Y, const std::string Text, unsigned long Shadow = 0);
 
 
 //////////////////////////////////////////////////////////////////////////////////////////
@@ -109,7 +109,7 @@ public:
 // Description:     Draws text to a bitmap aligned.
 // Arguments:       Bitmap, Position, Text.
 
-    void DrawAligned(GUIBitmap *Bitmap, int X, int Y, const std::string Text, int HAlign, int VAlign = Top, int maxWidth = 0, Uint32 Shadow = 0);
+    void DrawAligned(GUIBitmap *Bitmap, int X, int Y, const std::string Text, int HAlign, int VAlign = Top, int maxWidth = 0, unsigned long Shadow = 0);
 
 
 //////////////////////////////////////////////////////////////////////////////////////////
@@ -118,7 +118,7 @@ public:
 // Description:     Sets the current color.
 // Arguments:       Color.
 
-    void SetColor(Uint32 Color);
+    void SetColor(unsigned long Color);
 
 
 //////////////////////////////////////////////////////////////////////////////////////////
@@ -205,8 +205,8 @@ private:
     std::vector<FontColor >    m_ColorCache;
 
     int                m_FontHeight;
-    Uint32            m_MainColor;
-    Uint32            m_CurrentColor;
+    unsigned long            m_MainColor;
+    unsigned long            m_CurrentColor;
     GUIBitmap        *m_CurrentBitmap;
     std::string        m_Name;
     Character        m_Characters[256];
